README.txt
# CrossBridge
This game is about to cross an amount of people to the other side of the bridge. 
Every person has his own time it takes to cross.There is also a torch which is used so as people can cross 
and it can be hold by only one person each time.
To run the code write the times of the people separatin by " " (space) and write "0" (zero) to end the input line.
Code returns how many people are in each side, if it run correct StartSIde will be empty and all people will be on EndSide,
also returns the time it get to complete the proccess.